Work on the REST API

+++ app/Http/Controllers/api/BoardController.php
@@ -22,22 +22,22 @@ class BoardController extends Controller
     public function index()
     {
         try {
-        Log::Debug('BoardController@index');
+            Log::Debug('BoardController@index');
 
-        $elements = Board::all(); // SELECT * FROM boards
+            $elements = Board::all(); // SELECT * FROM boards
 
             return response()->json($elements, 200);
-        
+
         } catch (\Exception $e) {
 
             Log::Error('BoardController@index', ['message' => $e->getMessage()]);
-        $data = [
+            $data = [
                 'status' => 500,
                 'error' => 'Internal Server Error',
-        ];
+            ];
 
             return response()->json($data, 500);
-        }       
+        }
     }
 
     /**
@@ -46,19 +46,23 @@ class BoardController extends Controller
     public function show($id)
     {
         try {
-        Log::Debug("BoardController@show $id");
+            Log::Debug("BoardController@show $id");
 
-            $element = Board::findOrFail($id); // SELECT * FROM boards WHERE id = $id 
+            $element = Board::find($id); // SELECT * FROM boards WHERE id = $id
 
-            return response()->json($element, 200);
+            if ($element) {
+                return response()->json($element, 200);
+            } else {
+                return response()->json(['status' => 404, 'message' => "Board $id not found"], 404);
+            }
 
         } catch (\Exception $e) {
 
             Log::Error('BoardController@show', ['message' => $e->getMessage()]);
-        $data = [
+            $data = [
                 'status' => 500,
                 'error' => 'Internal Server Error',
-        ];
+            ];
 
             return response()->json($data, 500);
         }
@@ -70,130 +74,134 @@ class BoardController extends Controller
     public function store(Request $request)
     {
         try {
-        Log::Debug('BoardController@store');
+            Log::Debug('BoardController@store');
 
-        $validator = Validator::make($request->all(), [
-            "name" => 'required|string|max:128',
-            "description" => 'string|max:255',
-            "email" => 'required|string|max:128|email',
-            "favorite" => 'required|boolean',
-            "href" => 'string|max:255',
-            "image" => 'string|max:255',
-            "theme" => 'in:light,dark',
-			"lists" => ["string", "max:255", "regex:/\'(.+?)\'|\"(.+?)\"/"],
+            $validator = Validator::make($request->all(), [
+                "name" => 'required|string|max:128',
+                "description" => 'string|max:255',
+                "email" => 'required|string|max:128|email',
+                "favorite" => 'required|boolean',
+                "href" => 'string|max:255',
+                "image" => 'string|max:255',
+                "theme" => 'in:light,dark',
+                "lists" => ["string", "max:255", "regex:/\'(.+?)\'|\"(.+?)\"/"],
 
-        ]);
+            ]);
 
-        if ($validator->fails()) {
-            $data = [
-                'status' => 422,
-                'errors' => $validator->errors(),
-                'message' => 'Validation failed',
-            ];
-            Log::Debug('BoardController@store validation failed', $data);
+            if ($validator->fails()) {
+                $data = [
+                    'status' => 422,
+                    'errors' => $validator->errors(),
+                    'message' => 'Validation failed',
+                ];
+                Log::Debug('BoardController@store validation failed', $data);
 
-            return response()->json($data, 422);
-        }
+                return response()->json($data, 422);
+            }
 
-        $element = new Board;
-        $element->name = $request->name;
-        $element->description = $request->description;
-        $element->email = $request->email;
-        $element->favorite = $request->favorite;
-        $element->href = $request->href;
-        $element->image = $request->image;
-        $element->theme = $request->theme;
-		$element->lists = $request->lists;
+            $element = new Board;
+            $element->name = $request->name;
+            $element->description = $request->description;
+            $element->email = $request->email;
+            $element->favorite = $request->favorite;
+            $element->href = $request->href;
+            $element->image = $request->image;
+            $element->theme = $request->theme;
+            $element->lists = $request->lists;
 
-        $element->save();
-
+            $element->save();
 
             $data = [
                 'status' => 200,
-                'board' => $element
-            ];            
+                'board' => $element,
+            ];
             Log::Debug('BoardController@store saved in database', $data);
             return response()->json($element, 200);
 
         } catch (\Exception $e) {
 
             Log::Error('BoardController@store', ['message' => $e->getMessage()]);
-        $data = [
+            $data = [
                 'status' => 500,
                 'error' => 'Internal Server Error',
-        ];
+            ];
 
             return response()->json($data, 500);
-        }       
+        }
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, int $id)
+    public function update(Request $request, $id)
     {
         try {
-        Log::Debug("BoardController@update $id");
+            Log::Debug("BoardController@update $id");
 
-        $validator = Validator::make($request->all(), [
-            "name" => 'string|max:128',
-            "description" => 'string|max:255',
-            "email" => 'string|max:128|email',
-            "favorite" => 'boolean',
-            "href" => 'string|max:255',
-            "image" => 'string|max:255',
-            "theme" => 'in:light,dark',
-			"lists" => ["string", "max:255", "regex:/\'(.+?)\'|\"(.+?)\"/"],
+            $validator = Validator::make($request->all(), [
+                "name" => 'string|max:128',
+                "description" => 'string|max:255',
+                "email" => 'string|max:128|email',
+                "favorite" => 'boolean',
+                "href" => 'string|max:255',
+                "image" => 'string|max:255',
+                "theme" => 'in:light,dark',
+                "lists" => ["string", "max:255", "regex:/\'(.+?)\'|\"(.+?)\"/"],
 
-        ]);
-        if ($validator->fails()) {
-            $data = [
-                'status' => 422,
-                'errors' => $validator->errors(),
-                'message' => 'Validation failed',
-            ];
-            Log::Debug('BoardController@store validation failed', $data);
+            ]);
 
-            return response()->json($data, 422);
-        }
+            if ($validator->fails()) {
+                $data = [
+                    'status' => 422,
+                    'errors' => $validator->errors(),
+                    'message' => 'Validation failed',
+                ];
+                Log::Debug('BoardController@store validation failed', $data);
 
-            $element = Board::findOrFail($id);
-        if ($request->name) {
-            $element->name = $request->name;
-        }
-        if ($request->description) {
-            $element->description = $request->description;
-        }
-        if ($request->email) {
-            $element->email = $request->email;
-        }
-        if ($request->favorite) {
-            $element->favorite = $request->favorite;
-        }
-        if ($request->href) {
-            $element->href = $request->href;
-        }
-        if ($request->image) {
-            $element->image = $request->image;
-        }
-        if ($request->theme) {
-            $element->theme = $request->theme;
-        }
-		if ($request->lists) {
-			$element->lists = $request->lists;
-		}
+                return response()->json($data, 422);
+            }
 
-        $element->save();
+            $element = Board::find($id);
+            if (!$element) {
+                return response()->json(['status' => 404, 'message' => "Board $id not found"], 404);
+            }
+
+            if ($request->name) {
+                $element->name = $request->name;
+            }
+            if ($request->description) {
+                $element->description = $request->description;
+            }
+            if ($request->email) {
+                $element->email = $request->email;
+            }
+            if ($request->favorite) {
+                $element->favorite = $request->favorite;
+            }
+            if ($request->href) {
+                $element->href = $request->href;
+            }
+            if ($request->image) {
+                $element->image = $request->image;
+            }
+            if ($request->theme) {
+                $element->theme = $request->theme;
+            }
+            if ($request->lists) {
+                $element->lists = $request->lists;
+            }
+
+            $element->save();
 
             return response()->json($element, 200);
 
         } catch (\Exception $e) {
 
             Log::Error('BoardController@update', ['message' => $e->getMessage()]);
-        $data = [
+            $data = [
                 'status' => 500,
                 'error' => 'Internal Server Error',
-        ];
+            ];
 
             return response()->json($data, 500);
         }
@@ -205,9 +213,12 @@ class BoardController extends Controller
     public function destroy($id)
     {
         try {
-        Log::Debug("BoardController@delete $id");
+            Log::Debug("BoardController@delete $id");
 
-            $element = Board::findOrFail($id);
+            $element = Board::find($id);
+            if (!$element) {
+                return response()->json(['status' => 404, 'message' => "Board $id not found"], 404);
+            }
 
             $element->delete();
 
@@ -221,10 +232,10 @@ class BoardController extends Controller
         } catch (\Exception $e) {
 
             Log::Error('BoardController@destroy', ['message' => $e->getMessage()]);
-        $data = [
+            $data = [
                 'status' => 500,
                 'error' => 'Internal Server Error',
-        ];
+            ];
 
             return response()->json($data, 500);
         }
